+++ src/index.ts
@@ -1,5 +1,6 @@
 export * from './core/client';
 export * from './core/platform-client';
+export * from './core/request-error';
 export * from './types/items';
 export * from './types/base';
 export * from './types/execution';